Update documentation and rename module file
Update the CHANGELOG.md and README.md files with changes from the set of
functionality related to the roll ability scores epic for the gem.
Modify a comment in the roll_ability_scores file, and rename the
following file to try and reduce confusion between the
NerdDice::Configuration class and the configuration and configure
methods in the NerdDice module.

Old:
lib/nerd_dice/class_methods/configuration.rb
New:
lib/nerd_dice/class_methods/configure.rb

Completes NerdDice backlog items:
* Ensure comments are up-to-date before releasing 0.3
* Update CHANGELOG and README for 0.3 functionality
* Add method to roll ability scores #7
