+++ lib/nerd_dice/class_methods.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "nerd_dice/class_methods/configuration"
+require "nerd_dice/class_methods/configure"
 require "nerd_dice/class_methods/refresh_seed"
 require "nerd_dice/class_methods/execute_die_roll"
 require "nerd_dice/class_methods/total_dice"
@@ -15,7 +15,8 @@ require "nerd_dice/class_methods/total_ability_scores"
 # files that provide the class method functionality.
 #
 # PUBLIC METHODS
-# NerdDice.configuration => ./class_methods/configuration.rb
+# NerdDice.configure => ./class_methods/configure.rb
+# NerdDice.configuration => ./class_methods/configure.rb
 # NerdDice.refresh_seed => ./class_methods/refresh_seed.rb
 # NerdDice.execute_die_roll => ./class_methods/execute_die_roll.rb
 # NerdDice.total_dice => ./class_methods/total_dice.rb
