+++ lib/nerd_dice/class_methods/configure.rb
